+++ tests/__init__.py
@@ -1,6 +1,10 @@
 import os.path
 
 
+def path(filename):
+    return os.path.join('tests', 'fixtures', filename)
+
+
 def read(filename, mode='rt', encoding=None):
-    with open(os.path.join('tests', 'fixtures', filename), mode, encoding=encoding) as f:
+    with open(path(filename), mode, encoding=encoding) as f:
         return f.read()